64-bit: fix types to build for 64-bit targets
Some API within Audio HAL are inconsistant, switching from size_t
to uint32_t, preventing to switch to 64-bit.

This patch restores the consistency of the whole chain by switching
to size_t.

Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-4991
Reviewed-on: https://android.intel.com/343035
Tested-by: Esclafer, CatherineX <catherinex.esclafer@intel.com>

+++ audio_conversion/src/AudioRemapper.cpp
@@ -154,7 +154,7 @@ status_t AudioRemapper::convertChannelsPolicyInStereo(const void *src,
                                                       size_t *outFrames)
 {
     const type *srcTyped = static_cast<const type *>(src);
-    uint32_t frames = 0;
+    size_t frames = 0;
     uint32_t srcChannels = mSsSrc.getChannelCount();
 
     struct Stereo
